playground/timer_mailing/requirements.txt
aiogram>=3.0.0
python-dotenv>=1.0.0
apscheduler>=3.10.10